float to double

+++ src/main/java/activators/Activator.java
@@ -1,6 +1,6 @@
 package main.java.activators;
 
 public interface Activator {
-    public float Activate(float value);
-    public float Derivative(float value);
+    public double Activate(double value);
+    public double Derivative(double value);
 }
